fixing maze stuff

+++ PsyNaut/PsyNaut/Sources/MXMazeGenerator.h
@@ -5,14 +5,14 @@ typedef void (^block)(bool **);
 @interface MXMazeGenerator : NSObject
 {
 @private
-    uint _width;
-    uint _height;
+    NSUInteger _width;
+    NSUInteger _height;
     bool **_maze;
     NSMutableArray *_moves;
     CGPoint _start;
     block _response;
 }
 
-- (id)initWithRow:(int)row andCol:(int)col withStartingPoint:(CGPoint)position;
+- (id)initWithRow:(NSUInteger)row andCol:(NSUInteger)col withStartingPoint:(CGPoint)position;
 - (void)arrayMaze:(block)arrayMaze;
 @end
